pyproject.toml
[project]
name = "noiz"
version = "0.7.0"
authors = [
    { name = "Damian Kula", email = "heavelock@gmail.com" },
    { name = "Alexandre Kazantsev", email = "alexandre.kazantsev@storengy.com" },
    { name = "Arthur Dartois", email = "arthur.dartois@storengy.com" },
    { name = "Laure Duboeuf", email = "laure.duboeuf@external.storengy.com" },
]
readme = "README.rst"
license = { text = "CECILL-B" }
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: CeCILL-B Free Software License Agreement (CECILL-B)",
    "Topic :: Scientific/Engineering",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
requires-python = ">= 3.10, < 3.11"
dependencies = [
    "more-itertools >=8.11",
    "sqlalchemy >=1.4.0, <2.0",
    "click >=8.0.3",
    "environs >=9.3.5",
    "flask ==2.0.2",
    "flask-sqlalchemy ==2.5.1",
    "flask-migrate ==3.1.0",
    "loguru >=0.5.3",
    "matplotlib ==3.5.0",
    "numpy ==1.23.5",
    "obspy ==1.4.2",
    "pandas ==1.3.4",
    "psycopg2-binary ==2.9.2",
    "pydantic ~=1.8",
    "scipy ==1.15.3",
    "tqdm >=4.62.3",
    "toml>=0.10.2",
    "utm >=0.7.0",
    "dask[complete] ==2021.11.2",
    "pendulum ~=3.0",
    "werkzeug == 2.3.3",
]
[dependency-groups]
dev = [
    "pytest ==6.2.5",
    "pytest-check ==1.0.4",
    "pytest-cov ==3.0.0",
    "pytest-dependency ==0.5.1",
    "pytest-mock ==3.6.1",
    "mock==4.0.3",
    "mypy ==0.981",
    "sqlalchemy-stubs ==0.4",
    "types-python-dateutil ==2.8.2",
    "types-toml ==0.10.1",
]
docs = [
    "sphinx ~=7.0.0",
    "sphinx-autoapi ~=3.0.0",
    "sphinx-paramlinks ~=0.6.0",
    "sphinx-autodoc-typehints ~=1.24.0",
    "furo ~=2023.9.10",
]

lint = [
    "ruff ~=0.11.13",
    "sqlalchemy-stubs ==0.4",
    "types-python-dateutil ==2.8.2",
    "types-toml ==0.10.1",
    "pydantic ~=1.8",
    "doc8 ==0.10.1",
]

[project.urls]
Homepage = "https://noiz-group.gitlab.io/"
Documentation = "https://noiz-group.gitlab.io/noiz/"
Repository = "https://gitlab.com/noiz-group/noiz.git"
Changelog = "https://gitlab.com/noiz-group/noiz/-/blob/master/CHANGELOG.rst"

[project.scripts]
noiz = "noiz.cli:cli"

[build-system]
requires = ["uv_build>=0.7.6,<0.8.0"]
build-backend = "uv_build"

[tool.uv]
package = true

[tool.hatch.envs.test.scripts]
cov = 'SQLALCHEMY_WARN_20=1 pytest --cov=noiz'

[tool.hatch.envs.docs]
features = [
    "docs",
]

[tool.hatch.envs.docs.scripts]
build_html = [
    "sphinx-build -M html docs/ docs/_build ",
]


[tool.hatch.envs.lint.scripts]
run_doc8 = [
    "doc8 docs/content"
]


[tool.hatch.build.targets.sdist]
include = [
    "/src",
]

[tool.doc8]
max_line_length = 120
